refactored code based on the pr review.
